Add ExtensionCoreApp component and update dependencies
- Introduced the new `ExtensionCoreApp` component to encapsulate core application logic for the extension.
- Refactored `index.tsx` to utilize `ExtensionCoreApp`, improving structure and maintainability.
- Updated `yarn.lock` and `package.json` to include `@core/mpc` as a workspace dependency, ensuring proper module management.
